api/requirements.txt
alembic==1.17.2
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
pydot==4.0.1
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.1
pytest-env==1.2.0
python-dotenv==1.2.1
setuptools==80.9.0
SQLAlchemy==2.0.44
sqlalchemy_schemadisplay==2.0
typing_extensions==4.15.0
Werkzeug==3.1.3
